♻️[REFACTOR] Facade 패턴으로 변경 #15
